feat: redirect users to SSO's registration page
When a user is coming from a registration page, we want to redirect them
to the SSO's sign up page.

auth_entry is a query parameter that edx-platform passes to the
social-app-django's endpoint, and contains information about which form
a user is coming from, e.g. login or registration.

We use this value to determine when to redirect a user to the sign up
page, and use Auth0 screen-hint query string parameter to signify that
we want them to be redirected to the sign up page, instead of the login
one, which is the default.